requirements.txt
google_api_python_client==2.156.0
google_auth_oauthlib==1.0.0
Pillow==9.5.0
protobuf==5.29.2
